Move tests under test
