+++ tox.ini
@@ -23,7 +23,8 @@ commands =
     black --diff --check --exclude "/(\.eggs|\.git|\.tox|\.venv|\.build|dist|charmhelpers|mod)/" .
 
 [testenv:unit]
-commands = pytest -s scripts/upgrade-calico/test-upgrade-calico-cni.py
+setenv = PYTHONPATH={toxinidir}/scripts
+commands = pytest -s tests/unit/test-upgrade-calico-cni.py
 
 [testenv:clustering]
 commands = pytest -s tests/test-cluster.py
